backend/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt>=4.4.0,<4.5.0
psycopg2-binary>=2.8.0,<2.9.0
djoser==2.1.0
django-filter==2.4.0
python-decouple==3.4
django-autoslug==1.9.8
django-cors-headers==3.7.0
pytils==0.3
serpy==0.3.1
PyJWT==1.7.1
